streamlit/requirements.txt
streamlit==1.22.0
requests==2.28.2
Pillow==9.4.0 